done! all posts

+++ lib/views/posts/add_post.dart
@@ -240,8 +240,9 @@ class _AddPostState extends State<AddPost> {
                             "images": url,
                             "title": _titleController.text,
                             "locate": GeoPoint(pos!.latitude, pos!.longitude),
-                            "price": double.parse(_priceController.text),
-                            "surfaceArea": double.parse(_surfaceAreaController.text)
+                            "price": int.parse(_priceController.text),
+                            "surfaceArea":
+                                int.parse(_surfaceAreaController.text)
                           });
                           print(post);
                           Navigator.pushAndRemoveUntil(
@@ -249,10 +250,10 @@ class _AddPostState extends State<AddPost> {
                               MaterialPageRoute(
                                   builder: (context) => const Home()),
                               (route) => false);
-                          ScaffoldMessenger.of(context).showSnackBar(
-                               SnackBar(backgroundColor: Theme.of(context).colorScheme.primary,
-                                  content:
-                                      Text("Đã tạo bài đăng thành công!")));
+                          ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+                              backgroundColor:
+                                  Theme.of(context).colorScheme.primary,
+                              content: Text("Đã tạo bài đăng thành công!")));
                         }
                       }
                     } catch (e) {
